backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.29.0
starlette==0.41.3
motor==3.3.2
pymongo==4.6.1
PyJWT==2.8.0
bcrypt==4.1.2
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.3.2
httpx==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic[email]==2.6.4
anyio==4.3.0
